requirements.txt
pygame>=2.4.0
numpy>=1.24.0
scipy>=1.10.0
matplotlib>=3.7.0
pandas>=2.0.0
streamlit>=1.28.0
plotly>=5.15.0
PyYAML>=6.0
SALib>=1.4.7
reportlab>=4.0.0
tqdm>=4.65.0
